+++ DigitalSignal/__init__.py
@@ -1,4 +1,6 @@
-from .digital_signal import DigitalSignal, ShiftRegister
+from .digital_signal import DigitalSignal 
+from .shift_register import ShiftRegister
+from .helpers import E, PI, cconj
 
 # Make the class directly accessible at the package level
-__all__ = ['DigitalSignal', 'ShiftRegister', 'E', 'PI']
+__all__ = ['DigitalSignal', 'ShiftRegister', 'E', 'PI', 'cconj']
