+++ DigitalSignal/digital_signal.py
@@ -365,65 +365,3 @@ class DigitalSignal:
         clone.positive_indices = new_pos
         clone.negative_indices = new_neg
         return clone     
-
-def cconj(signal):
-    """
-    A helper function that calls the complex_conjugate method on a DigitalSignal object
-    """
-
-    try:
-        return signal.complex_conjugate() 
-    except AttributeError:
-        raise AttributeError
-
-
-class E():
-    def __init__(self, complex_const, ndigits=None):
-
-        if isinstance(complex_const, complex) and complex_const.real != 0:
-            raise TypeError("E can only take fully imagineary numbers (e.g. -2j not 3+4j)")
-
-        self.imag_const = complex_const.imag
-        self.ndigits = ndigits
-
-    def __call__(self, value):
-
-        z = cmath.cos(self.imag_const * value) + 1j * cmath.sin(self.imag_const * value)
-
-        if self.ndigits is None:
-            return z
-
-        return  complex(round(z.real, self.ndigits), round(z.imag, self.ndigits))
-
-PI = cmath.pi
-
-class ShiftRegister:
-    def __init__(self, length):
-        self.shift_size = length
-
-    def __call__(self, signal_length):
-        """
-        Generate a signal of the given length from this shift register object
-        """
-
-        # create a shift register of length 'shift_size' with a leading '1'
-        internal_signal = DigitalSignal([1]) + DigitalSignal()(-(self.shift_size-1))
-
-        # generate an output of length 'signal_length'
-        while len(internal_signal) < (signal_length):
-            
-            # modulus 2 add
-            new_value = internal_signal[0] ^ internal_signal[self.shift_size-1]
-            
-            # shift signal and append to end
-            internal_signal = internal_signal(-1)
-            internal_signal[0] = new_value
-
-        # convert 0s to -1s
-        internal_signal = 2 * internal_signal
-
-        internal_signal = internal_signal * (lambda a : a - 1) # subtract 1 from all explicit elements in the 
-
-        # reverse the explicit signal
-        return DigitalSignal(internal_signal[::-1])
-
